requirements.txt
absl-py==1.2.0
accelerate==0.24.1
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.8.1
aiosignal==1.2.0
alabaster==0.7.12
albumentations==1.4.0
alembic==1.8.1
altair==5.1.2
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.0.8
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
av==11.0.0
Babel==2.10.3
backcall==0.2.0
bagua-cuda113==0.9.0
base58==2.1.1
basicsr==1.4.2
beautifulsoup4==4.11.1
bitsandbytes==0.38.1
black==22.8.0
bleach==5.0.1
blessed==1.20.0
boto3==1.24.66
botocore==1.27.66
bravado==11.0.3
bravado-core==5.17.1
cachetools==5.2.0
certifi==2019.11.28
cffi==1.15.1
cfgv==3.3.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
cloudpickle==2.1.0
codecov==2.1.12
colorama==0.4.5
coloredlogs==15.0.1
colorlog==6.7.0
comet-ml==3.31.7
commonmark==0.9.1
ConfigArgParse==1.5.3
configobj==5.0.6
controlnet-aux==0.0.7
coverage==6.4.2
croniter==1.3.15
cryptography==37.0.4
cupy-cuda113==10.6.0
curio==1.5
cycler==0.11.0
Cython==3.0.8
databricks-cli==0.17.3
dateutils==0.6.12
dbus-python==1.2.16
debugpy==1.6.3
decorator==5.1.1
deepdiff==6.7.1
deepspeed==0.7.0
defusedxml==0.7.1
deprecation==2.1.0
diffusers==0.23.0
distlib==0.3.6
dlib==19.24.6
dnnlib @ https://github.com/podgorskiy/dnnlib/releases/download/0.0.1/dnnlib-0.0.1-py3-none-any.whl#sha256=e789afb6365198754f3eedd7687d2c321f6a2daf8009eb94e46fb8b90217361d
docker==5.0.3
docker-pycreds==0.4.0
docstring-parser==0.14.1
docutils==0.19
dulwich==0.20.45
easydict==1.12
einops==0.6.1
embreex==2.17.7.post3
entrypoints==0.4
envlight @ git+https://github.com/ashawkey/envlight.git@ef492c03711c87287549a0283ee51199f45cbea4
everett==3.0.0
executing==1.0.0
facexlib==0.3.0
fairscale==0.4.6
fastapi==0.1.17
fastjsonschema==2.16.1
fastrlock==0.8
ffmpy==0.3.1
filelock==3.8.0
filterpy==1.4.5
fire==0.4.0
Flask==2.2.2
flatbuffers==2.0.7
fonttools==4.37.1
fqdn==1.5.1
frozenlist==1.3.1
fsspec==2023.10.0
ftfy==6.1.1
future==0.18.2
gcsfs==2022.5.0
gevent==21.12.0
gfpgan==1.3.8
gitdb==4.0.9
GitPython==3.1.27
glfw==2.2.0
google-api-core==2.8.2
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
gorilla==0.4.0
gradio==4.4.0
gradio_client==0.7.0
graphviz==0.8.4
greenlet==1.1.3
grpcio==1.48.1
grpcio-tools==1.48.1
gunicorn==20.1.0
gym==0.24.1
gym-notices==0.0.8
h11==0.13.0
hivemind==1.0.1
hjson==3.1.0
horovod==0.25.0
httpcore==1.0.2
httpx==0.25.1
huggingface-hub==0.19.4
humanfriendly==10.0
hydra-core==1.2.0
identify==2.5.3
idna==2.8
imageio==2.32.0
imageio-ffmpeg==0.4.9
imagesize==1.4.1
importlib-metadata==6.8.0
importlib-resources==5.9.0
iniconfig==1.1.1
inquirer==3.1.3
insightface==0.7.3
ipykernel==6.15.2
ipyparallel==8.4.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
isoduration==20.11.0
itsdangerous==2.1.2
jaxtyping==0.2.19
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
jsonargparse==4.12.0
jsonpointer==2.3
jsonref==0.2
jsonschema==4.15.0
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kiwisolver==1.2.0
kornia==0.7.0
lazy_loader==0.3
legacy==0.1.7
libigl==2.4.1
lightning==2.0.0
lightning-cloud==0.5.54
lightning-utilities==0.9.0
llvmlite==0.41.0
lmdb==1.4.1
lxml==4.9.1
Mako==1.2.2
mapbox-earcut==1.0.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.6
mediapipe==0.10.9
mistune==2.0.4
mlflow==1.28.0
monotonic==1.6
mpmath==1.2.1
mrcfile==1.5.2
msgpack==1.0.4
multiaddr==0.0.9
multidict==6.0.2
mvdream @ file:///apdcephfs/private_haolinliu/4dfy/extern/MVDream
mxnet==1.6.0
mypy==0.971
mypy-extensions==0.4.3
nbclient==0.6.7
nbconvert==7.0.0
nbformat==5.4.0
neptune-client==0.16.3
nerfacc @ git+https://github.com/KAIR-BAIR/nerfacc.git@d84cdf3afd7dcfc42150e0f0506db58a5ce62812
nest-asyncio==1.5.5
netaddr==0.8.0
networkx==2.8.6
ninja==1.10.2.3
nodeenv==1.7.0
notebook==6.4.12
numba==0.58.0
numpy==1.23.0
nvdiffrast @ git+https://github.com/NVlabs/nvdiffrast.git@c5caf7bdb8a2448acc491a9faa47753972edd380
nvidia-dali-cuda110==1.16.1
nvidia-ml-py3==7.352.0
oauthlib==3.2.0
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.12.1
onnxruntime-gpu==1.17.1
open-clip-torch==2.24.0
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
OpenEXR==1.3.9
opt-einsum==3.3.0
ordered-set==4.1.0
orjson==3.9.10
outcome==1.2.0
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parallel-ssh==2.8.0
parso==0.8.3
pathspec==0.10.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
pluggy==1.0.0
plyfile==1.0.3
pre-commit==2.20.0
prefetch-generator==1.0.1
prettytable==3.10.0
prometheus-client==0.14.1
prometheus-flask-exporter==0.20.3
promise==2.3
prompt-toolkit==3.0.31
protobuf==4.25.3
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py-cpuinfo==8.0.0
pyaml==21.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycollada==0.7.2
pycparser==2.21
pycuda==2024.1.2
pydantic==2.5.1
pydantic_core==2.14.3
pydub==0.25.1
pygame==2.1.0
Pygments==2.13.0
PyGObject==3.36.0
PyJWT==2.4.0
PyMCubes==0.1.4
pymultihash==0.8.2
PyOpenGL==3.1.5
pyparsing==3.0.9
pyrsistent==0.18.1
pysdf==0.1.9
pyspng==0.1.1
pytest==7.0.1
pytest-asyncio==0.19.0
pytest-benchmark==3.4.1
pytest-cov==3.0.0
pytest-forked==1.4.0
pytest-rerunfailures==10.2
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytools==2024.1.13
pytorch-lightning==2.1.2
pytz==2022.2.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==23.2.1
qtconsole==5.3.2
QtPy==2.2.0
qudida==0.0.4
querystring-parser==1.2.4
readchar==4.0.5
regex==2023.10.3
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
rfc3339-validator==0.1.4
rfc3987==1.3.8
rich==12.5.1
rsa==4.9
Rtree==1.1.0
s3transfer==0.6.0
safetensors==0.4.0
scikit-image==0.21.0
scikit-learn==1.0.1
scikit-optimize==0.9.0
scipy==1.10.1
semantic-version==2.10.0
Send2Trash==1.8.0
sentencepiece==0.1.99
sentry-sdk==1.9.7
setproctitle==1.3.2
setuptools-rust==1.5.1
shapely==2.0.2
shellingham==1.5.4
shortuuid==1.0.9
simplejson==3.17.6
six==1.14.0
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
sounddevice==0.4.6
soupsieve==2.3.2.post1
Sphinx==5.1.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.40
sqlparse==0.4.2
ssh-python==0.10.0
ssh2-python==1.0.0
stack-data==0.5.0
starlette==0.19.1
starsessions==1.3.0
svg.path==6.3
swagger-spec-validator==2.7.6
sympy==1.11.1
tabulate==0.8.10
taming-transformers-rom1504==0.0.6
tb-nightly==2.11.0a20220906
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
terminado==0.15.0
test-tube==0.7.5
testpath==0.6.0
threadpoolctl==3.1.0
tifffile==2022.8.12
timm==0.9.10
tinycss2==1.1.1
tinycudann @ git+https://github.com/NVlabs/tiny-cuda-nn.git@c12645dfbac707ccd82907a209edff9b258f1e3d#subdirectory=bindings/torch
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.0
torch==1.12.1+cu113
torchmetrics==0.9.2
torchvision==0.13.1+cu113
tornado==6.2
tqdm==4.64.1
traitlets==5.3.0
transformers==4.28.1
trimesh==4.0.4
trio==0.21.0
typeguard==4.1.5
typer==0.9.0
typing_extensions==4.8.0
uri-template==1.2.0
urllib3==1.26.12
uvicorn==0.18.2
uvloop==0.16.0
varint==1.0.2
virtualenv==20.16.4
wandb==0.13.1
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.0
websockets==11.0.3
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wrapt==1.14.1
wurlitzer==3.0.2
xatlas==0.0.8
xxhash==3.0.0
yapf==0.40.2
yarl==1.8.1
zipp==3.8.1
zope.event==4.5.0
zope.interface==5.4.0
